+++ internal/civisibility/utils/telemetry/telemetry_count.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016 Datadog, Inc.
 
-package utils
+package telemetry
 
 import "gopkg.in/DataDog/dd-trace-go.v1/internal/telemetry"
